+++ flow-typed/store.js
@@ -2,3 +2,4 @@ declare type ParsedQueryResponses = {
 	successes: Array<QueryState>,
 	errors: Array<QueryState>,
 };
+declare type QueryFetcher = (queries: Array<Query>): Promise<ParsedQueryResponses>